refactor: split PHAR bridge
